backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
mysql-connector-python
requests
pydantic